docs: Switch to date-based releases
I usually make small, atomic changes that are available as soon as
they're pushed to GitHub, so the semantic versioning doesn't make too
much sense. From a user's perspective, almost everything is a breaking
change, and from the code perspective, it's a little fuzzy since it's
just a drop in theme, with the only interface being perhaps the config
file.